feat: integrate agent branch + dynamic seeders + TS singleton reconciliation
Cherry-pick integration (from agent branch → main):
- 49 traits, 28 enums, 60 models, 50 controllers, 8 middleware, 5 services
- 676 views, 160 seeders, 1,110 assets, ~870 new files
- 87 factories, ~290 tests, Python exporters/importers, Azure deploy configs
- PHPStan Level 3: 0 errors

Dynamic seeder profiles:
- Rewrite DatabaseSeeder.php with SEED_PROFILE env var (minimal|standard|full)
- Add SEED_ONLY / SEED_SKIP env vars for fine-grained control
- Create config/seeders.php with profile definitions
- Add SEED_PROFILE documentation to .env.example

Seeder backup protection:
- Sync 6 new seeders to _inc/.seeders/ and .backup/database/seeders/
- Add DO_NOT_MODIFY.md to both backup directories
- All 3 seeder locations now at 190 files

TypeScript singleton reconciliation:
- Create globals.d.ts with full type declarations for window.ERPGuard/ERPUtils/ERPBootstrap
- Add AUXILIARY warnings to TS core files (not runtime versions)
- Protect build pipeline: esm-to-iife.cjs excludes core by default (--include-core flag)
- tsc: 0 errors

+++ _inc/laravel/app/Models/Individuals/EmployeeAttendance.php
@@ -10,6 +10,22 @@ use Illuminate\Database\Eloquent\{
     Model,
     Relations\BelongsTo
 };
+/**
+ * @property \Illuminate\Support\Carbon|string|null $clock_in
+ * @property \Illuminate\Support\Carbon|string|null $clock_out
+ * @property mixed $created_by
+ * @property float|int|null $early_arrival_count
+ * @property string|null $early_leaving
+ * @property float|int|null $early_leaving_count
+ * @property float|int|null $late_count
+ * @property float|int|null $total_rest
+ * @property float|int|null $total_work
+ * @property \Illuminate\Support\Carbon|null $date
+ * @property mixed $late
+
+ * @property mixed $overtime
+ * @property int|null $status
+ */
 
 class EmployeeAttendance extends Model
 {
@@ -69,25 +85,25 @@ class EmployeeAttendance extends Model
     protected static function booted(): void
     {
         static::creating(function (self $model): void {
-            if (!$model->{AC::COL_TT_RST}) $model->{AC::COL_TT_RST} = '00:00:00';
+            if (!$model->{AC::COL_TT_RST}) $model->{AC::COL_TT_RST} = '00:00:00'; // @phpstan-ignore assign.propertyType
         });
 
         static::saving(function (self $model): void {
-            $model->status = AttendanceStatus::normalize($model->status ?? null)->value;
+            $model->status = (int)AttendanceStatus::normalize($model->status ?? null)->value;
 
             $model->{AC::COL_ERL_AV_CT}  = max(0, (int) ($model->{AC::COL_ERL_AV_CT} ?? 0));
             $model->{AC::COL_LT_CT}      = max(0, (int) ($model->{AC::COL_LT_CT} ?? 0));
             $model->{AC::COL_ERL_LV_CT}  = max(0, (int) ($model->{AC::COL_ERL_LV_CT} ?? 0));
             $model->{AC::COL_OVT_CT}     = max(0, (int) ($model->{AC::COL_OVT_CT} ?? 0));
 
-            $model->{AC::COL_TT_RST} = $model->normalizeDuration($model->{AC::COL_TT_RST} ?? '00:00:00') ?? '00:00:00';
+            $model->{AC::COL_TT_RST} = $model->normalizeDuration($model->{AC::COL_TT_RST} ?? '00:00:00') ?? '00:00:00'; // @phpstan-ignore assign.propertyType
 
             $model->ensureBefore(AC::COL_ERL_ARV, AC::COL_CLK_IN, AC::COL_ERL_AV_CT);
             $model->ensureAfter('late', AC::COL_CLK_IN, AC::COL_LT_CT);
             $model->ensureBefore(AC::COL_ERL_LV, AC::COL_CLK_OUT, AC::COL_ERL_LV_CT);
             $model->ensureAfter('overtime', AC::COL_CLK_OUT, AC::COL_OVT_CT);
 
-            $model->{AC::COL_TT_WRK} = $model->normalizeDuration(
+            $model->{AC::COL_TT_WRK} = $model->normalizeDuration( // @phpstan-ignore assign.propertyType
                 $model->{AC::COL_TT_WRK} ?: $model->calculateTotalWork()
             ) ?? '00:00:00';
         });
